backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
drf-nested-routers==0.93.4

# Production Server
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Task Queue
celery==5.3.4
redis==5.0.1

# LLM Integration
google-generativeai==0.3.2
openai==1.3.7

# File Processing
PyPDF2==3.0.1
pdfminer.six==20221105
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3

# Image Processing
opencv-python==4.8.1.78
scikit-image==0.21.0
Pillow==10.1.0

# Data Analysis
numpy==1.25.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Search and NLP
elasticsearch==8.11.0
sentence-transformers==2.2.2
nltk==3.8.1

# AWS S3
boto3==1.34.0
django-storages==1.14.2

# Environment and Configuration
python-decouple==3.8
django-environ==0.11.2

# API Documentation
drf-yasg==1.21.7

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Security
django-ratelimit==4.1.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.1.0 